refactor: split footnote plugins into per-mode packages on @vivliostyle/internal
Split the monolithic footnotes plugin into three published packages
(@vivliostyle/remark-footnote-{dpub,gcpm,pandoc}); vfm's footnotes plugin
becomes a thin switch over them. dpub and gcpm share the Properties-or-Factory
element customizer through @vivliostyle/internal (ElementFactory, buildElement,
HastPropertiesSchema, TagAwareH), so no separate remark-footnote-customizer
package is introduced; each package's public FootnoteFactory is aliased from
internal's generic ElementFactory.

Per-mode footnote unit tests move into their packages (vfm footnote coverage
narrows to switch logic: vfm 214 -> 183 tests, +31 across the new packages).

This squashed re-application of #251, rebased onto the table-cell/internal
branch, sources the customizer machinery from @vivliostyle/internal instead of
the standalone remark-footnote-customizer package #251 originally introduced.
