+++ packages/remark-footnote-pandoc/package.json
@@ -1,0 +1,40 @@
+{
+  "name": "@vivliostyle/remark-footnote-pandoc",
+  "version": "0.0.0",
+  "description": "Pandoc-style endnote rewriter used by VFM. Converts mdast-util-to-hast's default endnote markup into a `<section role=\"doc-endnotes\">` block with renumbering and duplicate-call backlinks.",
+  "license": "Apache-2.0",
+  "engines": {
+    "node": ">= 22"
+  },
+  "type": "module",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "files": [
+    "lib"
+  ],
+  "publishConfig": {
+    "access": "public"
+  },
+  "scripts": {
+    "build": "shx rm -rf lib && tsc",
+    "test": "tsc --project tests/tsconfig.json && vitest run",
+    "format": "prettier --write --ignore-path ../../.gitignore --ignore-path ../../.prettierignore .",
+    "lint": "eslint --fix ."
+  },
+  "dependencies": {
+    "@types/hast": "^2.0.0",
+    "@types/unist": "^2.0.0",
+    "hast-util-select": "^5.0.5",
+    "hastscript": "^7.2.0",
+    "unified": "^9.2.0",
+    "unist-builder": "^3.0.1",
+    "valibot": "1.2.0"
+  },
+  "devDependencies": {
+    "rehype-format": "^3.1.0",
+    "rehype-stringify": "^8.0.0",
+    "remark-footnotes": "^2.0.0",
+    "remark-parse": "^8.0.2",
+    "remark-rehype": "^8.1.0"
+  }
+}
