+++ packages/vfm/package.json
@@ -52,6 +52,9 @@
     "@types/mdast": "^3.0.0",
     "@types/unist": "^2.0.0",
     "@vivliostyle/internal": "workspace:*",
+    "@vivliostyle/remark-footnote-dpub": "workspace:*",
+    "@vivliostyle/remark-footnote-gcpm": "workspace:*",
+    "@vivliostyle/remark-footnote-pandoc": "workspace:*",
     "@vivliostyle/remark-ruby": "workspace:*",
     "@vivliostyle/remark-sectionize": "workspace:*",
     "debug": "^4.4.3",
