+++ packages/vfm/src/plugins/footnotes.ts
@@ -18,307 +18,42 @@
  *  `- mdast-util-to-hast@11.3.0      <-- Handler type & all() (direct import)
  */
 
-import type * as hast from 'hast';
-import { selectAll } from 'hast-util-select';
-import { h } from 'hastscript';
 import {
-  type Handler as ToHastHandler,
-  all as convertToHast,
-} from 'mdast-util-to-hast';
+  createDpubFootnotePlugin,
+  DpubFootnoteOptionsSchema,
+  DpubYamlFootnoteOptionsSchema,
+  type DpubFootnoteOptions,
+} from '@vivliostyle/remark-footnote-dpub';
+import {
+  createGcpmFootnotePlugin,
+  GcpmFootnoteOptionsSchema,
+  GcpmYamlFootnoteOptionsSchema,
+  type GcpmFootnoteOptions,
+} from '@vivliostyle/remark-footnote-gcpm';
+import {
+  createPandocFootnotePlugin,
+  PandocFootnoteOptionsSchema,
+} from '@vivliostyle/remark-footnote-pandoc';
+import type { Handler as ToHastHandler } from 'mdast-util-to-hast';
 import footnotes from 'remark-footnotes';
-import type * as unist from 'unist';
-import { u } from 'unist-builder';
+import type unified from 'unified';
 import * as v from 'valibot';
-import { mergePlugins, partial } from '../utils.js';
-import {
-  buildElement,
-  HastPropertiesSchema,
-  type ElementFactory,
-  type TagAwareH,
-} from '@vivliostyle/internal';
+import { partial } from '../utils.js';
 
-type ElementWithProps = hast.Element & {
-  properties: NonNullable<hast.Element['properties']>;
-};
-
-/**
- * @see https://github.com/syntax-tree/mdast-util-to-hast/blob/10.2.0/lib/footer.js#L55-L66
- */
-const endnoteAreaSelector = 'div.footnotes';
-type EndnoteArea = hast.Element & {
-  tagName: 'div';
-  properties: { className: ['footnotes'] };
-};
-const selectEndnoteAreas = (tree: hast.Root) =>
-  selectAll(endnoteAreaSelector, tree) as EndnoteArea[];
-
-/**
- * @see https://github.com/syntax-tree/mdast-util-to-hast/blob/10.2.0/lib/footer.js#L43-L48
- */
-const endnoteElementSelector = `${endnoteAreaSelector} ol li[id^="fn-"]`;
-type EndnoteElement = hast.Element & {
-  tagName: 'li';
-  properties: { id: `fn-${string}` };
-};
-const selectEndnoteElements = (tree: hast.Root) =>
-  selectAll(endnoteElementSelector, tree) as EndnoteElement[];
-
-/**
- * @see https://github.com/syntax-tree/mdast-util-to-hast/blob/10.2.0/lib/handlers/footnote-reference.js#L15-L19
- */
-const endnoteCallSelector =
-  'sup[id^="fnref-"]:has(a:only-child[href^="#fn-"].footnote-ref)';
-type EndnoteCall = hast.Element & {
-  tagName: 'sup';
-  properties: { id: `fnref-${string}` };
-  children: [
-    hast.Element & {
-      tagName: 'a';
-      properties: {
-        href: `#${string}`;
-        className: ['footnote-ref'];
-      };
-    },
-  ];
-};
-const selectEndnoteCalls = (tree: hast.Root) =>
-  selectAll(endnoteCallSelector, tree) as EndnoteCall[];
-
-/**
- * @see https://github.com/syntax-tree/mdast-util-to-hast/blob/10.2.0/lib/footer.js#L29-L34
- */
-const endnoteBackReferenceSelector = 'a.footnote-backref';
-type EndnoteBackReference = hast.Element & {
-  tagName: 'a';
-  properties: { className: ['footnote-backref'] };
-};
-const selectEndnoteBackReferences = (parent: hast.Element) =>
-  selectAll(endnoteBackReferenceSelector, parent) as EndnoteBackReference[];
-
-/**
- * Return a copy of `array` with only `key`'s values permuted to follow
- * `compareFn`'s sort order; every other field stays in its original
- * position.  Used to reshuffle one field across fixed slots without
- * disturbing surrounding entries.
- */
-function withSortedField<T extends Record<string, unknown>, K extends keyof T>(
-  array: readonly T[],
-  key: K,
-  compareFn: (a: T, b: T) => number,
-): T[] {
-  const sortedValues = array.toSorted(compareFn);
-  return array.map(
-    (item, i) => ({ ...item, [key]: sortedValues[i]![key] }) as T,
-  );
-}
-
-/**
- * Create paired Pandoc transformer plugins that share duplicate-reference
- * state.  The first plugin rewrites endnote calls; the second rewrites
- * endnote areas and adds extra backlinks for duplicate references.
- *
- * This style is named "pandoc" because the original implementation
- * (commit 146abea) was described as Pandoc-like. The closest versions
- * are 2.7 and 2.8-2.13 (which added the variant selector U+FE0E to
- * the backlink U+21A9). No Pandoc version is an exact match (checked
- * up to 3.9.0, the latest release as of 2026-03). Endnote body text
- * here is not wrapped in <p>.
- *
- *   $ printf 'Text[^1].\n\n[^1]: Footnote.' | docker run --rm -i pandoc/core:2.7 -f markdown -t html
- *   <p>Text<a href="#fn1" class="footnote-ref" id="fnref1" role="doc-noteref"><sup>1</sup></a>.</p>
- *   <section class="footnotes" role="doc-endnotes">
- *   <hr />
- *   <ol>
- *   <li id="fn1" role="doc-endnote"><p>Footnote.<a href="#fnref1" class="footnote-back" role="doc-backlink">↩</a></p></li>
- *   </ol>
- *   </section>
- *
- * Duplicate references to the same endnote definition are handled
- * differently from both Pandoc and GFM (tested via remark-gfm@4.0.1,
- * the latest as of 2026-03), though closer to the latter.
- */
-const createPandocTransformers = () => {
-  // Shared between the two plugins: how many duplicate calls each
-  // refIndex has.  Populated by the calls transformer, read by the
-  // areas transformer.
-  const dupCountByRefIndex = new Map<number, number>();
-  // identifier to refIndex, established by the calls transformer from
-  // document-order of references.  The areas transformer uses this to
-  // rebind endnote bodies regardless of the order mdast-util-to-hast
-  // emitted them (issue #129: table cells are walked in reverse by
-  // mdast-util-to-hast <12.1.1, producing a reversed endnote list).
-  const identifierToRefIndex = new Map<string, number>();
-
-  const endnoteCallsToPandoc = () => (tree: unist.Node) => {
-    // Reset closure-level maps so each document processed by a shared
-    // processor instance starts from a clean slate.
-    identifierToRefIndex.clear();
-    dupCountByRefIndex.clear();
-
-    const calls = selectEndnoteCalls(tree as hast.Root);
-
-    // Build a stable mapping from definition identifier to endnote number.
-    // Duplicate references to the same definition reuse that number.
-    const dupCount = new Map<string, number>();
-    let counter = 0;
-
-    calls.forEach((call) => {
-      const sup: ElementWithProps = call;
-      const [anchor]: [ElementWithProps] = call.children;
-
-      const href = String(anchor.properties.href);
-      const identifier = href.replace(/^#fn-/, '');
-
-      let refIndex = identifierToRefIndex.get(identifier);
-      let callId: string;
-
-      if (refIndex === undefined) {
-        refIndex = ++counter;
-        identifierToRefIndex.set(identifier, refIndex);
-        callId = `fnref${refIndex}`;
-      } else {
-        const dup = (dupCount.get(identifier) ?? 0) + 1;
-        dupCount.set(identifier, dup);
-        dupCountByRefIndex.set(refIndex, dup);
-        callId = `fnref${refIndex}-${dup}`;
-      }
-
-      // Mutate sup into <a class="footnote-ref">
-      sup.tagName = 'a';
-      sup.properties = {
-        id: callId,
-        href: `#fn${refIndex}`,
-        className: ['footnote-ref'],
-        role: 'doc-noteref',
-      };
-
-      // Mutate anchor into <sup>N</sup>
-      anchor.tagName = 'sup';
-      anchor.properties = {};
-      anchor.children = [u('text', `${refIndex}`)];
-    });
-  };
-
-  const endnoteAreasToPandoc = () => (tree: unist.Node) => {
-    const root = tree as hast.Root;
-
-    // must be called before mutating area.tagName
-    const endnoteElements = selectEndnoteElements(root);
-
-    selectEndnoteAreas(root).forEach((area: ElementWithProps) => {
-      area.tagName = 'section';
-      area.properties.role = 'doc-endnotes';
-    });
-
-    // Resolve each endnote element's true refIndex from the identifier
-    // assigned during the calls pass, then relabel and reorder by refIndex.
-    // When the upstream emitter is already fixed, this pass is idempotent
-    // (the order and ids converge to the same values either way).
-    const resolved = endnoteElements
-      .map((elem: ElementWithProps) => {
-        const originalId = String(elem.properties.id ?? '');
-        const identifier = originalId.replace(/^fn-/, '');
-        const refIndex = identifierToRefIndex.get(identifier);
-        return { elem, refIndex };
-      })
-      .filter(
-        (x): x is { elem: ElementWithProps; refIndex: number } =>
-          x.refIndex !== undefined,
-      );
-
-    resolved.forEach(({ elem, refIndex }) => {
-      elem.properties.id = `fn${refIndex}`;
-      elem.properties.role = 'doc-endnote';
-
-      // Back reference is expected to be placed at the end of contents
-      // @see https://github.com/syntax-tree/mdast-util-to-hast/blob/10.2.0/lib/footer.js#L41
-      selectEndnoteBackReferences(elem).forEach((backref: ElementWithProps) => {
-        backref.properties.href = `#fnref${refIndex}`;
-        backref.properties.className = ['footnote-back'];
-        backref.properties.role = 'doc-backlink';
-      });
-
-      // Add extra backlinks for duplicate references
-      const dups = dupCountByRefIndex.get(refIndex);
-      if (dups) {
-        for (let d = 1; d <= dups; d++) {
-          elem.children.push(
-            h(
-              'a',
-              {
-                href: `#fnref${refIndex}-${d}`,
-                className: ['footnote-back'],
-                role: 'doc-backlink',
-              },
-              '↩',
-            ),
-          );
-        }
-      }
-    });
-
-    // Reorder <li> within each <ol> to match refIndex ascending so the
-    // rendered list counter matches the id.  Non-<li> siblings keep their
-    // relative positions among <li> slots.
-    const refIndexByElem = new Map<hast.Element, number>(
-      resolved.map((r) => [r.elem, r.refIndex]),
-    );
-    selectAll('section[role="doc-endnotes"] ol', root).forEach((ol) => {
-      withSortedField(
-        ol.children.flatMap((child, i) =>
-          child.type === 'element' &&
-          child.tagName === 'li' &&
-          refIndexByElem.has(child)
-            ? [
-                {
-                  li: child,
-                  fn: (li: hast.Element) => {
-                    ol.children[i] = li;
-                  },
-                },
-              ]
-            : [],
-        ),
-        'li',
-        ({ li: a }, { li: b }) =>
-          (refIndexByElem.get(a) ?? 0) - (refIndexByElem.get(b) ?? 0),
-      ).forEach(({ fn, li }) => {
-        fn(li);
-      });
-    });
-  };
-
-  return [endnoteCallsToPandoc, endnoteAreasToPandoc] as const;
-};
-
-/**
- * Factory that customizes a footnote element. Footnote-specific alias of the
- * generic {@link ElementFactory}, preserved for backward compatibility.
- */
-export type FootnoteFactory<
-  TTag extends string,
-  TProps,
-  TChildren extends hast.ElementContent[] = hast.ElementContent[],
-> = ElementFactory<TTag, TProps, TChildren>;
-
-/** Backlink element placed at the head of a DPUB footnote body. */
-type DpubBacklink = hast.Element & {
-  tagName: 'a';
-  children: [
-    hast.Element & {
-      tagName: 'sup';
-      children: [hast.Text & { value: `${number}` }];
-    },
-  ];
-  properties: {
-    href: `#fnref${number}`;
-    className: ['footnote-back'];
-    role: 'doc-backlink';
-  };
-};
-
-/** Children tuple passed to DPUB body factory / buildElement. */
-export type DpubBodyChildren = [DpubBacklink, ...hast.ElementContent[]];
+// Re-export factory types so vfm consumers (and the index.ts re-exports)
+// keep their existing import paths.
+export type {
+  DpubCallFactory,
+  DpubBodyFactory,
+  DpubBodyChildren,
+  DpubCallChildren,
+  FootnoteFactory as DpubFootnoteFactory,
+} from '@vivliostyle/remark-footnote-dpub';
+export type {
+  GcpmBodyFactory,
+  GcpmDuplicatedCallFactory,
+  FootnoteFactory as GcpmFootnoteFactory,
+} from '@vivliostyle/remark-footnote-gcpm';
 
 /**
  * - `"pandoc"`: endnote section at document end (default).
@@ -328,82 +63,10 @@ export type DpubBodyChildren = [DpubBacklink, ...hast.ElementContent[]];
  */
 export type FootnoteMode = 'pandoc' | 'dpub' | 'gcpm';
 
-/** Sup element wrapping the reference number in a DPUB noteref. */
-type DpubCallSup = hast.Element & {
-  tagName: 'sup';
-  children: [hast.Text & { value: `${number}` }];
-};
-
-/** Children tuple passed to DPUB call factory / buildElement. */
-export type DpubCallChildren = [DpubCallSup];
-
-export type DpubCallFactory = FootnoteFactory<
-  'a',
-  {
-    id: `fnref${number}`;
-    href: `#fn${number}`;
-    class: 'footnote-ref';
-    role: 'doc-noteref';
-  },
-  DpubCallChildren
->;
-
-export type DpubBodyFactory = FootnoteFactory<
-  'aside',
-  { id: `fn${number}`; class: 'footnote'; role: 'doc-footnote' },
-  DpubBodyChildren
->;
-
-export type GcpmBodyFactory = FootnoteFactory<
-  'span',
-  { id: `fn-${string}`; role: 'doc-footnote' }
->;
-
-export type GcpmDuplicatedCallFactory = FootnoteFactory<
-  'a',
-  {
-    href: `#fn-${string}`;
-    class: 'footnote-duplicated-call';
-    role: 'doc-noteref';
-  }
->;
-
 export const FootnoteModeSchema = v.union([
   v.literal('pandoc'),
   v.literal('dpub'),
   v.literal('gcpm'),
-]);
-
-const dpubCallSchema = v.union([
-  HastPropertiesSchema,
-  v.pipe(
-    v.function() as v.GenericSchema<DpubCallFactory>,
-    v.metadata({ typeString: 'DpubCallFactory' }),
-  ),
-]);
-
-const dpubBodySchema = v.union([
-  HastPropertiesSchema,
-  v.pipe(
-    v.function() as v.GenericSchema<DpubBodyFactory>,
-    v.metadata({ typeString: 'DpubBodyFactory' }),
-  ),
-]);
-
-const gcpmBodySchema = v.union([
-  HastPropertiesSchema,
-  v.pipe(
-    v.function() as v.GenericSchema<GcpmBodyFactory>,
-    v.metadata({ typeString: 'GcpmBodyFactory' }),
-  ),
-]);
-
-const gcpmDuplicatedCallSchema = v.union([
-  HastPropertiesSchema,
-  v.pipe(
-    v.function() as v.GenericSchema<GcpmDuplicatedCallFactory>,
-    v.metadata({ typeString: 'GcpmDuplicatedCallFactory' }),
-  ),
 ]);
 
 export const FootnoteOptionsSchema = v.object({
@@ -411,17 +74,9 @@ export const FootnoteOptionsSchema = v.object({
     v.union([
       FootnoteModeSchema,
       v.variant('mode', [
-        v.object({ mode: v.literal('pandoc') }),
-        v.object({
-          mode: v.literal('dpub'),
-          call: v.optional(dpubCallSchema),
-          body: v.optional(dpubBodySchema),
-        }),
-        v.object({
-          mode: v.literal('gcpm'),
-          body: v.optional(gcpmBodySchema),
-          duplicatedCall: v.optional(gcpmDuplicatedCallSchema),
-        }),
+        PandocFootnoteOptionsSchema,
+        DpubFootnoteOptionsSchema,
+        GcpmFootnoteOptionsSchema,
       ]),
     ]),
   ),
@@ -430,26 +85,18 @@ export const FootnoteOptionsSchema = v.object({
 export type FootnoteOptions = v.InferInput<typeof FootnoteOptionsSchema>;
 
 /**
- * YAML-safe variant of {@link FootnoteOptionsSchema}: `call`, `body`, and
- * `duplicatedCall` accept `hast.Properties` only, since YAML cannot
- * represent JavaScript functions.
+ * YAML-safe variant of {@link FootnoteOptionsSchema}: customizer fields
+ * accept `hast.Properties` only, since YAML cannot represent JavaScript
+ * functions.
  */
 export const YamlFootnoteOptionsSchema = v.object({
   footnote: v.optional(
     v.union([
       FootnoteModeSchema,
       v.variant('mode', [
-        v.object({ mode: v.literal('pandoc') }),
-        v.object({
-          mode: v.literal('dpub'),
-          call: v.optional(HastPropertiesSchema),
-          body: v.optional(HastPropertiesSchema),
-        }),
-        v.object({
-          mode: v.literal('gcpm'),
-          body: v.optional(HastPropertiesSchema),
-          duplicatedCall: v.optional(HastPropertiesSchema),
-        }),
+        PandocFootnoteOptionsSchema,
+        DpubYamlFootnoteOptionsSchema,
+        GcpmYamlFootnoteOptionsSchema,
       ]),
     ]),
   ),
@@ -459,495 +106,10 @@ export type YamlFootnoteOptions = v.InferInput<
   typeof YamlFootnoteOptionsSchema
 >;
 
-type BuildFootnote = (
-  id: `fn-${string}`,
-  children: hast.ElementContent[],
-) => hast.Element;
-
-const createBuildFootnote =
-  (factory: GcpmBodyFactory): BuildFootnote =>
-  (id, children) => {
-    const result = factory(
-      h as TagAwareH,
-      { id, role: 'doc-footnote' as const },
-      children,
-    );
-    result.tagName = 'span';
-    return result;
-  };
-
-type BuildDuplicatedCall = (targetId: `fn-${string}`) => hast.Element;
-
-const createBuildDuplicatedCall =
-  (
-    customizer: hast.Properties | GcpmDuplicatedCallFactory | undefined,
-  ): BuildDuplicatedCall =>
-  (targetId) =>
-    buildElement(
-      'a',
-      {
-        href: `#${targetId}`,
-        class: 'footnote-duplicated-call',
-        role: 'doc-noteref',
-      } as {
-        href: `#fn-${string}`;
-        class: 'footnote-duplicated-call';
-        role: 'doc-noteref';
-      },
-      [],
-      customizer,
-    );
-
-const createInlineFootnoteHandler =
-  (buildFootnote: BuildFootnote): ToHastHandler =>
-  (ctx, node) => {
-    let no = 1;
-    while (String(no) in ctx.footnoteById) {
-      no++;
-    }
-    const identifier = String(no);
-    ctx.footnoteById[identifier] = {
-      type: 'footnoteDefinition',
-      identifier,
-      children: [{ type: 'paragraph', children: node.children }],
-      position: node.position,
-    };
-    return buildFootnote(`fn-${identifier}`, convertToHast(ctx, node));
-  };
-
-const createFootnoteReferenceHandler = (
-  buildFootnote: BuildFootnote,
-  buildDuplicatedCall: BuildDuplicatedCall,
-): ToHastHandler => {
-  const seen = new Set<string>();
-  return (ctx, node) => {
-    const identifier = String(node.identifier);
-    const def = ctx.footnoteById[identifier.toUpperCase()];
-    if (!def) {
-      return null;
-    }
-
-    if (seen.has(identifier)) {
-      return buildDuplicatedCall(`fn-${identifier}`);
-    }
-
-    seen.add(identifier);
-    return buildFootnote(
-      `fn-${identifier}`,
-      convertToHast(
-        ctx,
-        // Unwrap single-paragraph definitions to produce inline content
-        // (matches tight list-item behavior in footer.js).
-        def.children[0]?.type === 'paragraph' ? def.children[0] : def,
-      ),
-    );
-  };
-};
-
-/**
- * Create mdast-to-hast handlers and hast transformers for footnote processing.
- *
- * When `footnote` is `"pandoc"` or unset (default):
- * - `toHastHandlers` is empty; the default mdast-util-to-hast handlers
- *   generate endnote sections.
- * - `hastTransformers` rewrites the endnote markup to Pandoc format.
- *
- * When `footnote` is `"gcpm"`:
- * - `toHastHandlers` provides custom `footnoteReference` and `footnote`
- *   handlers that produce `<span>` footnote elements at the call site.
- *   A second reference to the same definition produces an
- *   `<a class="footnote-duplicated-call">` instead of another `<span>`.
- *
- * When `footnote` is `"dpub"`:
- * - `toHastHandlers` produces `<a role="doc-noteref">` calls, queues
- *   `<aside role="doc-footnote">` elements, and emits placeholders at
- *   footnote definition positions.
- * - `hastTransformers` replaces placeholders with the queued asides.
- *   Inline footnotes (no definition position) are inserted after their
- *   nearest non-flow-container ancestor.
- *
- * @param options Footnote options.
- * @returns `toHastHandlers` for remark-rehype and `hastTransformers` as
- *   unified plugins to be spread into the pipeline.
- */
-const createDpubFootnoteReferenceHandler =
-  (
-    pending: Map<string, hast.Element>,
-    nextIndex: () => number,
-    assigned: Map<string, { refIndex: number; dupCount: number }>,
-    callCustomizer: hast.Properties | DpubCallFactory | undefined,
-    bodyCustomizer: hast.Properties | DpubBodyFactory | undefined,
-  ): ToHastHandler =>
-  (ctx, node) => {
-    const identifier = String(node.identifier);
-    const def = ctx.footnoteById[identifier.toUpperCase()];
-    if (!def) {
-      return null;
-    }
-
-    const existing = assigned.get(identifier);
-
-    if (existing) {
-      // Duplicate reference: reuse the same footnote number
-      const refIndex = existing.refIndex;
-      const fnId = `fn${refIndex}` as `fn${number}`;
-      existing.dupCount++;
-      const callId =
-        `fnref${refIndex}-${existing.dupCount}` as `fnref${number}`;
-
-      return buildElement(
-        'a',
-        {
-          id: callId,
-          href: `#${fnId}`,
-          class: 'footnote-ref' as const,
-          role: 'doc-noteref' as const,
-        },
-        [h('sup', u('text', `${refIndex}`))] as DpubCallChildren,
-        callCustomizer,
-      );
-    }
-
-    // First reference: assign a new number and create the aside
-    const refIndex = nextIndex();
-    assigned.set(identifier, { refIndex, dupCount: 0 });
-    const callId = `fnref${refIndex}` as `fnref${number}`;
-    const fnId = `fn${refIndex}` as `fn${number}`;
-
-    const backlink: DpubBacklink = h(
-      'a',
-      { href: `#${callId}`, class: 'footnote-back', role: 'doc-backlink' },
-      h('sup', u('text', `${refIndex}`)),
-    ) as DpubBacklink;
-
-    const bodyChildren: DpubBodyChildren = [
-      backlink,
-      ...convertToHast(
-        ctx,
-        def.children[0]?.type === 'paragraph' ? def.children[0] : def,
-      ),
-    ];
-
-    pending.set(
-      identifier,
-      buildElement(
-        'aside',
-        { id: fnId, class: 'footnote' as const, role: 'doc-footnote' as const },
-        bodyChildren,
-        bodyCustomizer,
-      ),
-    );
-
-    return buildElement(
-      'a',
-      {
-        id: callId,
-        href: `#${fnId}`,
-        class: 'footnote-ref' as const,
-        role: 'doc-noteref' as const,
-      },
-      [h('sup', u('text', `${refIndex}`))] as DpubCallChildren,
-      callCustomizer,
-    );
-  };
-
-const createDpubInlineFootnoteHandler =
-  (
-    inlinePending: Map<string, hast.Element>,
-    nextIndex: () => number,
-    callCustomizer: hast.Properties | DpubCallFactory | undefined,
-    bodyCustomizer: hast.Properties | DpubBodyFactory | undefined,
-  ): ToHastHandler =>
-  (ctx, node) => {
-    let no = 1;
-    while (String(no) in ctx.footnoteById) {
-      no++;
-    }
-    const identifier = String(no);
-    ctx.footnoteById[identifier] = {
-      type: 'footnoteDefinition',
-      identifier,
-      children: [{ type: 'paragraph', children: node.children }],
-      position: node.position,
-    };
-
-    const refIndex = nextIndex();
-    const callId = `fnref${refIndex}` as `fnref${number}`;
-    const fnId = `fn${refIndex}` as `fn${number}`;
-
-    const backlink: DpubBacklink = h(
-      'a',
-      { href: `#${callId}`, class: 'footnote-back', role: 'doc-backlink' },
-      h('sup', u('text', `${refIndex}`)),
-    ) as DpubBacklink;
-
-    const bodyChildren: DpubBodyChildren = [
-      backlink,
-      ...convertToHast(ctx, node),
-    ];
-
-    inlinePending.set(
-      callId,
-      buildElement(
-        'aside',
-        { id: fnId, class: 'footnote' as const, role: 'doc-footnote' as const },
-        bodyChildren,
-        bodyCustomizer,
-      ),
-    );
-
-    return buildElement(
-      'a',
-      {
-        id: callId,
-        href: `#${fnId}`,
-        class: 'footnote-ref' as const,
-        role: 'doc-noteref' as const,
-      },
-      [h('sup', u('text', `${refIndex}`))] as DpubCallChildren,
-      callCustomizer,
-    );
-  };
-
-/**
- * Emit a placeholder element at the footnote definition position.
- * The default mdast-util-to-hast handler ignores footnoteDefinition nodes;
- * this override preserves the position in the hast tree so the transformer
- * can replace the placeholder with the queued aside.
- */
-const createDpubFootnoteDefinitionHandler =
-  (): ToHastHandler => (_ctx, node) => {
-    const identifier = String(node.identifier);
-    return h('div', { dataFootnotePlaceholder: identifier });
-  };
-
-/**
- * Replace definition-position placeholders with queued asides, and insert
- * inline-footnote asides after their nearest non-flow-container ancestor.
- */
-const createReplaceDpubPlaceholders =
-  (
-    pending: Map<string, hast.Element>,
-    inlinePending: Map<string, hast.Element>,
-  ) =>
-  () =>
-  (tree: unist.Node) => {
-    const root = tree as hast.Root;
-    const placed = new Set<string>();
-
-    // Phase 1: Replace placeholders with asides from reference footnotes
-    function walkAndReplace(parent: hast.Element | hast.Root) {
-      const newChildren: (hast.RootContent | hast.ElementContent)[] = [];
-      for (const child of parent.children) {
-        if (
-          child.type === 'element' &&
-          child.tagName === 'div' &&
-          child.properties?.dataFootnotePlaceholder
-        ) {
-          const id = String(child.properties.dataFootnotePlaceholder);
-          const aside = pending.get(id);
-          if (aside) {
-            newChildren.push(aside);
-            placed.add(id);
-          }
-          // Unreferenced definition: drop placeholder
-          continue;
-        }
-        newChildren.push(child);
-        if (child.type === 'element') {
-          walkAndReplace(child);
-        }
-      }
-      parent.children = newChildren;
-    }
-
-    walkAndReplace(root);
-
-    // Phase 2: Insert inline footnote asides after nearest non-flow-container
-    if (inlinePending.size > 0) {
-      const remaining = new Set(inlinePending.keys());
-
-      const containsCall = (
-        node: hast.RootContent | hast.ElementContent,
-      ): string[] => {
-        if (remaining.size === 0 || node.type !== 'element') {
-          return [];
-        }
-        const id = node.properties?.id;
-        return [
-          ...(typeof id === 'string' && remaining.has(id) ? [id] : []),
-          ...node.children.flatMap((child) => containsCall(child)),
-        ];
-      };
-
-      const processParent = (parent: hast.Element | hast.Root) => {
-        const newChildren: (hast.RootContent | hast.ElementContent)[] = [];
-        for (const child of parent.children) {
-          newChildren.push(child);
-          if (child.type !== 'element') {
-            continue;
-          }
-          if (flowContainerTagNames.has(child.tagName)) {
-            processParent(child);
-          } else {
-            for (const callId of containsCall(child)) {
-              const aside = inlinePending.get(callId);
-              if (aside) {
-                newChildren.push(aside);
-                remaining.delete(callId);
-              }
-            }
-          }
-        }
-        parent.children = newChildren;
-      };
-
-      processParent(root);
-
-      for (const callId of remaining) {
-        const aside = inlinePending.get(callId);
-        if (aside) {
-          root.children.push(aside);
-        }
-      }
-    }
-
-    // Phase 3: Renumber refIndex based on DOM order of calls.
-    //
-    // The footnote handlers assign refIndex at handler-invocation time,
-    // which mdast-util-to-hast <12.1.1 walks in reverse inside table
-    // cells (issue #129).  Now that the tree is fully built, rewrite
-    // every fnref${N}[-${dup}] / fn${N} using a mapping derived from
-    // the document order of call sites.  When the upstream walker is
-    // already fixed this mapping is the identity, making the pass a
-    // no-op.
-    const oldToNewRefIndex = new Map<number, number>();
-    let newCounter = 0;
-
-    (selectAll('a.footnote-ref', root) as ElementWithProps[]).forEach(
-      (anchor) => {
-        const callId = String(anchor.properties.id ?? '');
-        const match = /^fnref(\d+)(?:-(\d+))?$/.exec(callId);
-        if (!match) return;
-        const oldRefIndex = Number(match[1]);
-        const dupSuffix = match[2];
-
-        if (!oldToNewRefIndex.has(oldRefIndex)) {
-          newCounter++;
-          oldToNewRefIndex.set(oldRefIndex, newCounter);
-        }
-        const newRefIndex = oldToNewRefIndex.get(oldRefIndex) as number;
-
-        anchor.properties.id =
-          dupSuffix === undefined
-            ? `fnref${newRefIndex}`
-            : `fnref${newRefIndex}-${dupSuffix}`;
-        anchor.properties.href = `#fn${newRefIndex}`;
-        const sup = anchor.children.find(
-          (c): c is hast.Element => c.type === 'element' && c.tagName === 'sup',
-        );
-        if (sup && sup.children[0]?.type === 'text') {
-          sup.children[0].value = String(newRefIndex);
-        }
-      },
-    );
-
-    (
-      selectAll('aside[role="doc-footnote"]', root) as ElementWithProps[]
-    ).forEach((aside) => {
-      const id = String(aside.properties.id ?? '');
-      const match = /^fn(\d+)$/.exec(id);
-      if (!match) return;
-      const oldRefIndex = Number(match[1]);
-      const newRefIndex = oldToNewRefIndex.get(oldRefIndex);
-      if (newRefIndex === undefined) return;
-
-      aside.properties.id = `fn${newRefIndex}`;
-
-      const backlink = aside.children.find((c): c is hast.Element => {
-        if (c.type !== 'element' || c.tagName !== 'a') return false;
-        const className = c.properties?.className;
-        return (
-          Array.isArray(className) &&
-          (className as unknown[]).includes('footnote-back')
-        );
-      });
-      if (backlink) {
-        const oldHref = String(backlink.properties?.href ?? '');
-        const backMatch = /^#fnref(\d+)(-\d+)?$/.exec(oldHref);
-        if (backMatch && backlink.properties) {
-          const suffix = backMatch[2] ?? '';
-          backlink.properties.href = `#fnref${newRefIndex}${suffix}`;
-        }
-        const backSup = backlink.children.find(
-          (c): c is hast.Element => c.type === 'element' && c.tagName === 'sup',
-        );
-        if (backSup && backSup.children[0]?.type === 'text') {
-          backSup.children[0].value = String(newRefIndex);
-        }
-      }
-    });
-  };
-
-/**
- * Elements that accept flow content as children, into which aside
- * insertion recurses to find a more precise insertion point.
- * For all other elements the aside is placed immediately after the
- * element in its parent.
- *
- * Transparent elements are included because this set is only consulted
- * from within a flow container context, so they inherit flow content
- * model from their parent.
- * @see https://html.spec.whatwg.org/multipage/dom.html#flow-content
- * @see https://html.spec.whatwg.org/multipage/dom.html#transparent-content-models
- */
-const flowContainerTagNames = new Set([
-  'address',
-  'article',
-  'aside',
-  'blockquote',
-  'body',
-  'dd',
-  'details',
-  'dialog',
-  'div',
-  'fieldset',
-  'figcaption',
-  'figure',
-  'footer',
-  'form',
-  'header',
-  'li',
-  'main',
-  'nav',
-  'search',
-  'section',
-  'td',
-  'th',
-  // Transparent content model
-  'a',
-  'audio',
-  'canvas',
-  'del',
-  'ins',
-  'map',
-  'object',
-  'slot',
-  'video',
-]);
-
 type ResolvedOption =
   | { mode: 'pandoc' }
-  | {
-      mode: 'dpub';
-      call?: hast.Properties | DpubCallFactory | undefined;
-      body?: hast.Properties | DpubBodyFactory | undefined;
-    }
-  | {
-      mode: 'gcpm';
-      body?: hast.Properties | GcpmBodyFactory | undefined;
-      duplicatedCall?: hast.Properties | GcpmDuplicatedCallFactory | undefined;
-    };
+  | DpubFootnoteOptions
+  | GcpmFootnoteOptions;
 
 const resolveOption = (opt: FootnoteOptions['footnote']): ResolvedOption => {
   if (!opt) {
@@ -972,79 +134,23 @@ export type FootnoteToHastHandlers = {
   footnote: ToHastHandler | undefined;
 };
 
-export const createFootnotePlugin = (options?: FootnoteOptions) => {
+export const createFootnotePlugin = (
+  options?: FootnoteOptions,
+): {
+  toHastHandlers: FootnoteToHastHandlers;
+  hastTransformer: unified.Plugin;
+} => {
   const resolved = resolveOption(options?.footnote);
 
   if (resolved.mode === 'pandoc') {
-    const [endnoteCallsToPandoc, endnoteAreasToPandoc] =
-      createPandocTransformers();
-    return {
-      toHastHandlers: {
-        footnoteDefinition: undefined,
-        footnoteReference: undefined,
-        footnote: undefined,
-      },
-      hastTransformer: mergePlugins(endnoteCallsToPandoc, endnoteAreasToPandoc),
-    };
+    return createPandocFootnotePlugin();
   }
 
   if (resolved.mode === 'dpub') {
-    const pending = new Map<string, hast.Element>();
-    const inlinePending = new Map<string, hast.Element>();
-    const assigned = new Map<string, { refIndex: number; dupCount: number }>();
-    let counter = 0;
-    const nextIndex = () => ++counter;
-    return {
-      toHastHandlers: {
-        footnoteDefinition: createDpubFootnoteDefinitionHandler(),
-        footnoteReference: createDpubFootnoteReferenceHandler(
-          pending,
-          nextIndex,
-          assigned,
-          resolved.call,
-          resolved.body,
-        ),
-        footnote: createDpubInlineFootnoteHandler(
-          inlinePending,
-          nextIndex,
-          resolved.call,
-          resolved.body,
-        ),
-      },
-      hastTransformer: createReplaceDpubPlaceholders(pending, inlinePending),
-    };
+    return createDpubFootnotePlugin(resolved);
   }
 
-  // gcpm
-  const body = resolved.body;
-  const buildFootnote = createBuildFootnote(
-    typeof body === 'function'
-      ? body
-      : typeof body === 'object'
-        ? (((hFn, props, children) =>
-            hFn('span', { ...props, ...body }, ...children)) as GcpmBodyFactory)
-        : (((hFn, props, children) =>
-            hFn(
-              'span',
-              { class: 'footnote', ...props },
-              ...children,
-            )) as GcpmBodyFactory),
-  );
-  const buildDuplicatedCall = createBuildDuplicatedCall(
-    resolved.duplicatedCall,
-  );
-
-  return {
-    toHastHandlers: {
-      footnoteDefinition: undefined,
-      footnoteReference: createFootnoteReferenceHandler(
-        buildFootnote,
-        buildDuplicatedCall,
-      ),
-      footnote: createInlineFootnoteHandler(buildFootnote),
-    },
-    hastTransformer: mergePlugins(),
-  };
+  return createGcpmFootnotePlugin(resolved);
 };
 
 /**
